feat: Push forward.

+++ SastImg/SastImg.Application/AlbumServices/CreateAlbum/CreateAlbumCommand.cs
@@ -9,7 +9,7 @@ namespace SastImg.Application.AlbumServices.CreateAlbum
     public sealed class CreateAlbumCommand(
         string title,
         string description,
-        long categoryId,
+        int categoryId,
         Accessibility accessibility,
         ClaimsPrincipal user
     ) : ICommandRequest<CreateAlbumDto>
