+++ SastImg/SastImg.Domain/CategoryEntity/CategoryId.cs
@@ -1,6 +1,6 @@
 ﻿namespace SastImg.Domain.CategoryEntity
 {
-    public readonly record struct CategoryId(long Value)
+    public readonly record struct CategoryId(int Value)
     {
         public override string ToString()
         {
